requirements.txt
streamlit>=1.32.0
python-dotenv
groq>=0.5.0
pdfplumber>=0.10.0
python-docx>=1.1.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
